add MonographDaoImpl, changed Publication to Monograph

+++ scientific-activity-parent/scientific-activity-common/src/main/resources/db.migration/V1__Create_monografs.sql
@@ -1,10 +1,10 @@
 CREATE TABLE monograph
 (
-    id          INT          NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    id          BIGINT       NOT NULL PRIMARY KEY AUTO_INCREMENT,
     type        VARCHAR(255) NOT NULL,
     title       VARCHAR(255) NOT NULL,
     edition     VARCHAR(255) NOT NULL,
-    author_id   INT          NOT NULL,
+    author_id   BIGINT       NOT NULL,
     printRun    INT          NOT NULL,
     pages       INT          NOT NULL,
     entryDate   DATETIME     NOT NULL
